+++ src/main/java/com/loopon/llm/infrastructure/gemini/GeminiPropertiesConfig.java
@@ -1,4 +1,4 @@
-package com.loopon.journey.infrastructure.llm;
+package com.loopon.llm.infrastructure.gemini;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
